refactor(colors): refactored according to the specifications of the team leader
