--- src/shared/colors/components/grid-color/styles.js
@@ -1,12 +1,0 @@
-import styled from "styled-components";
-
-const StyledGrid = styled.div`
-  display: grid;
-  max-width: 200px;
-  grid-template-columns: auto auto auto;
-  grid-template-rows: 1fr 1fr;
-  column-gap: 10px;
-  align-items: center;
-`;
-
-export { StyledGrid };
