+++ src/shared/colors/stories/system/styles.js
